+++ src/SmartTransactionsController.test.ts
@@ -460,6 +460,79 @@ describe('SmartTransactionsController', () => {
       } as NetworkState);
       expect(checkPollSpy).toHaveBeenCalled();
     });
+
+    it('calls "ensureUniqueSmartTransactions" on network change if it is a new chainId', () => {
+      const { smartTransactionsState } = smartTransactionsController.state;
+      const smartTransactionsForChainId = createStateAfterSuccess();
+      smartTransactionsForChainId.push({
+        // Duplicate a smart transaction with the same uuid.
+        ...smartTransactionsForChainId[0],
+        status: 'pending',
+      });
+      smartTransactionsController.update({
+        smartTransactionsState: {
+          ...smartTransactionsState,
+          smartTransactions: {
+            [ChainId.mainnet]:
+              smartTransactionsForChainId as SmartTransaction[],
+          },
+        },
+      });
+      smartTransactionsController.config.chainId = ChainId.sepolia;
+      networkListener({
+        providerConfig: {
+          chainId: ChainId.mainnet,
+          type: 'rpc',
+          ticker: 'ETH',
+        },
+        selectedNetworkClientId: 'networkClientId',
+        networkConfigurations: {},
+        networksMetadata: {},
+      } as NetworkState);
+      const uniqueSmartTransactionsForChainId =
+        smartTransactionsController.state.smartTransactionsState
+          .smartTransactions[ChainId.mainnet];
+      expect(uniqueSmartTransactionsForChainId).toHaveLength(1);
+      expect(uniqueSmartTransactionsForChainId).toStrictEqual([
+        smartTransactionsForChainId[0],
+      ]);
+    });
+
+    it('does not call "ensureUniqueSmartTransactions" on network change for the same chainId', () => {
+      const { smartTransactionsState } = smartTransactionsController.state;
+      const smartTransactionsForChainId = createStateAfterSuccess();
+      smartTransactionsForChainId.push({
+        // Duplicate a smart transaction with the same uuid.
+        ...smartTransactionsForChainId[0],
+        status: 'pending',
+      });
+      smartTransactionsController.update({
+        smartTransactionsState: {
+          ...smartTransactionsState,
+          smartTransactions: {
+            [ChainId.mainnet]:
+              smartTransactionsForChainId as SmartTransaction[],
+          },
+        },
+      });
+      networkListener({
+        providerConfig: {
+          chainId: ChainId.mainnet,
+          type: 'rpc',
+          ticker: 'ETH',
+        },
+        selectedNetworkClientId: 'networkClientId',
+        networkConfigurations: {},
+        networksMetadata: {},
+      } as NetworkState);
+      const currentSmartTransactionsForChainId =
+        smartTransactionsController.state.smartTransactionsState
+          .smartTransactions[ChainId.mainnet];
+      expect(currentSmartTransactionsForChainId).toHaveLength(2);
+      expect(currentSmartTransactionsForChainId).toStrictEqual(
+        smartTransactionsForChainId,
+      );
+    });
   });
 
   describe('checkPoll', () => {
@@ -718,7 +791,7 @@ describe('SmartTransactionsController', () => {
       });
       const pendingState = createStateAfterPending()[0];
       const pendingTransaction = { ...pendingState, history: [pendingState] };
-      expect(smartTransactionsController.state).toStrictEqual({
+      expect(smartTransactionsController.state).toMatchObject({
         smartTransactionsState: {
           smartTransactions: {
             [ChainId.mainnet]: [pendingTransaction],
@@ -770,7 +843,7 @@ describe('SmartTransactionsController', () => {
       });
       const successState = createStateAfterSuccess()[0];
       const successTransaction = { ...successState, history: [successState] };
-      expect(smartTransactionsController.state).toStrictEqual({
+      expect(smartTransactionsController.state).toMatchObject({
         smartTransactionsState: {
           smartTransactions: {
             [ChainId.mainnet]: [